pyproject.toml
[tool.poetry]
name = "ska-mid-itf"
version = "30.3.1rc1"
description = "Integration testing and deployment project for the Mid ITF"
authors = ["ATLAS"]
license = "BSD 3-Clause"
readme = "README.md"
packages = [{ include = "tests/integration" }]

[[tool.poetry.source]]
name = 'skao'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[[tool.poetry.source]]
name = "PyPI-public"
url = 'https://pypi.org/simple'

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
ska-tmc-cdm = ">=8.3.0"
pytest-json-report = "^1.5.0"
pytest-repeat = "^0.9.1"
ska_ser_logging = "^0.4.1"
pyyaml = "^6.0.1"
ska-mid-itf-engineering-tools = "0.10.4-rc.3"
httpx = "^0.27.0"
h11 = "^0.14.0"
annotated-types = ">=0.6.0,<1.0.0"
black = "^24.3.0"
ska-telmodel = "^1.13.0"
ska-oso-scripting = "17.0.0"
astropy = "^6.0.0" 
pydantic = {extras = ["email"], version = "^2.12.5"}
ska-schemas = "^1.36.0"
psycopg = {extras = ["binary"], version = "^3.3.3"}
assertpy = "^1.1"
lxml = "^6.0.4"
diagrams = "^0.25.1"
mock = "^5.2.0"

[tool.poetry.group.docs.dependencies]
Sphinx = "^7.2.6"
sphinx-rtd-theme = "^1.3.0"
recommonmark = "^0.7.1"
markupsafe = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.12.0"
pylint = "^3.1.0"
pylint-junit = "^0.3.2"
pytest-cov = "^4.0.0"
flake8-docstrings = "^1.7.0"
flake8-rst-docstrings = "^0.3.0"
darglint = "^1.8.1"
mypy = "^1.8.0"
twine = "^5.0.0"
ipykernel = "^6.29.5"

[tool.poetry.group.ansible.dependencies]
ansible = "^9.3.0"
ansible-compat = "^25.1.0"
ansible-lint = { version = "^24.2.0", markers = "platform_system != 'Windows'" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-p no:warnings"
testpaths = ["tests/integration"]
log_cli_level = "INFO"
log_cli = "True"
junit_family = "xunit2"
python_classes = "!Test"
#bdd_features_base_dir = "tests"

[tool.build_sphinx]
source-dir = "docs/src"
build-dir = "docs/_build"
all_files = 1
builder = "html"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 99

[tool.flake8]
max-line-length = 99

[tool.black]
line_length = 99

[tool.poetry.scripts]
telescope_state_control = "scripts.telescope_control.state:main"
upload-to-confluence = "scripts.confluence.upload_to_confluence:main"